requirements.txt
azure-identity==1.2.0                                
azure-keyvault-secrets==4.0.0
confuse==1.0.0
pyOpenSSL==19.1.0   
logger==1.4 